app/modules/asignacion/ai_scaffold.py
"""
Módulo de scaffolding para integración con IA
TODO: Reemplazar placeholders con integraciones reales
"""
from typing import Optional, List, Dict, Any


# ============================================================================
# SCAFFOLDING: Speech-to-Text (Whisper / Google Speech)
# ============================================================================
async def transcribe_audio(audio_url: str) -> Optional[str]:
    """
    TODO IA: Enviar audio a modelo Whisper/Speech-to-Text
    
    Integración sugerida:
    - OpenAI Whisper API
    - Google Cloud Speech-to-Text
    - AWS Transcribe
    
    Ejemplo con OpenAI:
    ```python
    import openai
    response = openai.Audio.transcribe("whisper-1", audio_file)
    return response['text']
    ```
    """
    # Placeholder: retornar transcripción dummy
    return "El vehículo presenta falla en el sistema de frenos. Se escucha ruido metálico al frenar."


# ============================================================================
# SCAFFOLDING: Computer Vision para clasificación de imágenes
# ============================================================================
async def classify_image(image_url: str) -> Dict[str, Any]:
    """
    TODO IA: Enviar imagen a modelo de Computer Vision para clasificación
    
    Categorías esperadas: [batería, llanta, choque, motor, otros]
    
    Integración sugerida:
    - Google Cloud Vision API
    - AWS Rekognition
    - Modelo custom entrenado con TensorFlow/PyTorch
    
    Ejemplo con Google Vision:
    ```python
    from google.cloud import vision
    client = vision.ImageAnnotatorClient()
    # ... análisis de imagen
    return {'categoria': 'llanta', 'confianza': 0.95}
    ```
    """
    # Placeholder: retornar clasificación dummy
    return {
        'categoria': 'llanta',
        'confianza': 0.92,
        'etiquetas': ['pinchazo', 'rueda', 'emergencia'],
        'descripcion': "Se detecta posible daño en neumático"
    }


# ============================================================================
# SCAFFOLDING: Generación de resumen con IA
# ============================================================================
async def generate_summary(
    texto: Optional[str],
    transcripcion: Optional[str],
    clasificacion_imagen: Optional[Dict]
) -> Optional[str]:
    """
    TODO IA: Generar resumen consolidado usando LLM
    
    Integración sugerida:
    - OpenAI GPT-4 / GPT-3.5-turbo
    - Anthropic Claude
    - Modelo local fine-tuned
    
    Prompt sugerido:
    "Como asistente de emergencias vehiculares, genera un resumen conciso 
    de la situación basado en: descripción del usuario '{texto}', 
    transcripción de audio '{transcripcion}', y análisis de imagen '{clasificacion}'.
    Incluye: tipo de problema, urgencia estimada, y herramientas necesarias."
    """
    # Placeholder: retornar resumen dummy
    return "Emergencia: Pinchazo de llanta en vehículo sedan. Cliente varado en vía principal. Se requiere cambio de neumático urgente. Prioridad: Alta."


# ============================================================================
# SCAFFOLDING: Motor de asignación inteligente
# ============================================================================
def calculate_distance(
    lat1: float, lon1: float, 
    lat2: float, lon2: float
) -> float:
    """
    Calcular distancia entre dos coordenadas usando fórmula Haversine
    """
    from math import radians, sin, cos, sqrt, atan2
    
    R = 6371  # Radio de la Tierra en km
    
    lat1, lon1, lat2, lon2 = map(radians, [lat1, lon1, lat2, lon2])
    dlat = lat2 - lat1
    dlon = lon2 - lon1
    
    a = sin(dlat/2)**2 + cos(lat1) * cos(lat2) * sin(dlon/2)**2
    c = 2 * atan2(sqrt(a), sqrt(1-a))
    
    return R * c


async def smart_assign_taller(
    solicitud_id: int,
    solicitud_lat: float,
    solicitud_lon: float,
    tipo_problema: Optional[str],
    prioridad: Optional[str],
    talleres_disponibles: List[Dict]
) -> Optional[Dict]:
    """
    TODO IA: Motor de asignación que evalúe múltiples factores
    
    Factores a considerar:
    1. Distancia geográfica (Haversine) - peso: 40%
    2. Disponibilidad del taller - peso: 20%
    3. Capacidad (cantidad de técnicos disponibles) - peso: 15%
    4. Especialidad vs tipo de problema - peso: 15%
    5. Calificación histórica del taller - peso: 10%
    
    Algoritmo sugerido:
    - Score = Σ(factor_i * peso_i)
    - Seleccionar taller con mayor score
    - Si hay empate, priorizar por menor distancia
    
    Integración ML sugerida:
    - Modelo de ranking aprendido con historial de asignaciones exitosas
    - Reinforcement learning para optimizar asignaciones en tiempo real
    """
    if not talleres_disponibles:
        return None
    
    # Placeholder: filtrado básico por distancia
    resultados = []
    for taller in talleres_disponibles:
        if taller.get('disponible') and taller.get('latitud') and taller.get('longitud'):
            distancia = calculate_distance(
                solicitud_lat, solicitud_lon,
                taller['latitud'], taller['longitud']
            )
            resultados.append({
                **taller,
                'distancia_km': round(distancia, 2),
                'score': 100 - (distancia * 2)  # Score dummy inverso a distancia
            })
    
    # Ordenar por score descendente
    resultados.sort(key=lambda x: x['score'], reverse=True)
    
    return resultados[0] if resultados else None